Reorganized observation and action functions.
