+++ smarts/core/sensors.py
@@ -34,18 +34,17 @@ from smarts.core import coordinates
 from smarts.core import scenario
 from smarts.core import waypoints
 from smarts.core.coordinates import BoundingBox, Heading, HeadingMethods
-from smarts.core.events import Events, Collision
+from smarts.core.events import Events
 from smarts.core.lidar import Lidar
 from smarts.core.lidar_sensor_params import SensorParams
 from smarts.core.mission_planner import MissionPlanner
 from smarts.core.masks import RenderMasks
 from smarts.core.renderer import Renderer
 from smarts.core.scenario import MissionData, Via
-from smarts.core.utils import sequence
-from smarts.core.utils.math import squared_dist, vec_2d
+from smarts.core.utils.math import squared_dist
 from smarts.core.waypoints import Waypoint
-from smarts.sstudio.types import CutIn, UTurn, ZoneData
 from smarts.proto import observation_pb2
+from smarts.sstudio.types import CutIn, UTurn, ZoneData
 
 
 logger = logging.getLogger(__name__)
@@ -1280,149 +1279,3 @@ class ViaSensor(Sensor):
 
     def teardown(self):
         pass
-
-
-def fix_observation_size(obs_config: Dict, obs: Dict) -> Dict:
-    fixed_obs = {
-        agent_id: fix_agent_observation_size(obs_config, agent_obs)
-        for agent_id, agent_obs in obs.items()
-    }
-    return fixed_obs
-
-
-def fix_agent_observation_size(obs_config: Dict, obs: NamedTuple) -> Observation:
-
-    # Truncate/pad observation.events
-    collisions = sequence.truncate_pad_li(
-        obs.events.collisions,
-        obs_config["observation"]["events"]["collisions"],
-        Collision(),
-    )
-    events = obs.events._replace(collisions=collisions)
-
-    # Truncate/pad observation.ego_vehicle_state
-    yaw_rate = (sequence.truncate_pad_arr(obs.ego_vehicle_state.yaw_rate, 3, 0),)
-    route_vias = sequence.truncate_pad_li(
-        obs.ego_vehicle_state.mission.route_vias,
-        obs_config["observation"]["ego_vehicle_state"]["mission"]["route_vias"],
-        None,
-    )
-    via = (
-        sequence.truncate_pad_li(
-            obs.ego_vehicle_state.mission.via,
-            obs_config["observation"]["ego_vehicle_state"]["mission"]["via"],
-            Via(),
-        ),
-    )
-    mission = obs.ego_vehicle_state.mission._replace(
-        entry_tactic=None,  # EntryTactic removed from observation output
-        task=None,  # Task removed from observation output
-        via=via,
-        route_vias=route_vias,
-    )
-    linear_velocity = (
-        sequence.truncate_pad_arr(obs.ego_vehicle_state.linear_velocity, 3, 0),
-    )
-    angular_velocity = (
-        sequence.truncate_pad_arr(obs.ego_vehicle_state.angular_velocity, 3, 0),
-    )
-    linear_acceleration = (
-        sequence.truncate_pad_arr(obs.ego_vehicle_state.linear_acceleration, 3, 0),
-    )
-    angular_acceleration = (
-        sequence.truncate_pad_arr(obs.ego_vehicle_state.angular_acceleration, 3, 0),
-    )
-    linear_jerk = (sequence.truncate_pad_arr(obs.ego_vehicle_state.linear_jerk, 3, 0),)
-    angular_jerk = (
-        sequence.truncate_pad_arr(obs.ego_vehicle_state.angular_jerk, 3, 0),
-    )
-    ego_vehicle_state = obs.ego_vehicle_state._replace(
-        mission=mission,
-        linear_velocity=linear_velocity,
-        angular_velocity=angular_velocity,
-        linear_acceleration=linear_acceleration,
-        angular_acceleration=angular_acceleration,
-        linear_jerk=linear_jerk,
-        angular_jerk=angular_jerk,
-    )
-
-    # Truncate/pad observation.neighborhood_vehicle_states
-    neighborhood_vehicle_states = obs.neighborhood_vehicle_states or []
-    neighborhood_vehicle_states = (
-        sequence.truncate_pad_li(
-            neighborhood_vehicle_states,
-            obs_config["observation"]["neighborhood_vehicle_states"],
-            VehicleObservation(),
-        ),
-    )
-
-    # Truncate/pad observation.lidar_point_cloud
-    lidar_points = sequence.truncate_pad_li(
-        obs.lidar_point_cloud[0],
-        obs_config["observation"]["lidar_point_cloud"][0],
-        np.array([0, 0, 0], dtype=np.float32),
-    )
-    lidar_hits = sequence.truncate_pad_li(
-        obs.lidar_point_cloud[1],
-        obs_config["observation"]["lidar_point_cloud"][1],
-        np.array([0, 0, 0], dtype=np.float32),
-    )
-    lidar_ray = sequence.truncate_pad_li(
-        obs.lidar_point_cloud[2],
-        obs_config["observation"]["lidar_point_cloud"][2],
-        (np.array([0, 0, 0], dtype=np.float32), np.array([0, 0, 0], dtype=np.float32)),
-    )
-    lidar_point_cloud = (lidar_points, lidar_hits, lidar_ray)
-
-    # Truncate/pad observation.road_waypoints
-    if obs.road_waypoints:
-        lanes = {
-            k: sequence.truncate_pad_li(
-                v,
-                obs_config["observation"]["road_waypoints"]["lanes"],
-                Waypoint(),
-            )
-            for k, v in obs.road_waypoints.lanes.items()
-        }
-        route_waypoints = sequence.truncate_pad_li(
-            obs.road_waypoints.route_waypoints,
-            obs_config["observation"]["road_waypoints"]["route_waypoints"],
-            Waypoint(),
-        )
-        road_waypoints = RoadWaypoints(lanes=lanes, route_waypoints=route_waypoints)
-    else:
-        road_waypoints = obs.road_waypoints
-
-    # Truncate/pad observation.via_data
-    near_via_points = sequence.truncate_pad_li(
-        obs.via_data.near_via_points,
-        obs_config["observation"]["via_data"]["near_via_points"],
-        ViaPoint(),
-    )
-    hit_via_points = sequence.truncate_pad_li(
-        obs.via_data.hit_via_points,
-        obs_config["observation"]["via_data"]["hit_via_points"],
-        ViaPoint(),
-    )
-    via_data = Vias(near_via_points=near_via_points, hit_via_points=hit_via_points)
-
-    # Fixed-size observation
-    fixed_obs = Observation(
-        events=events,
-        ego_vehicle_state=ego_vehicle_state,
-        neighborhood_vehicle_states=neighborhood_vehicle_states,
-        waypoint_paths=sequence.truncate_pad_li_2d(
-            obs.waypoint_paths,
-            obs_config["observation"]["waypoint_paths"],
-            ([], Waypoint()),
-        ),
-        distance_travelled=obs.distance_travelled,
-        lidar_point_cloud=lidar_point_cloud,
-        drivable_area_grid_map=obs.drivable_area_grid_map,
-        occupancy_grid_map=obs.occupancy_grid_map,
-        top_down_rgb=obs.top_down_rgb,
-        road_waypoints=road_waypoints,
-        via_data=via_data,
-    )
-
-    return fixed_obs
